requirements.txt
reflex==0.6.6.post3
python-decouple
openai
